requirements.txt.txt
python-telegram-bot>=20.0
openai>=1.0.0
pandas>=1.0.0
aiohttp>=3.0.0  # python-telegram-bot的依赖
httpx>=0.25.0  # python-telegram-bot 20.0+的HTTP客户端
asyncio>=3.0  # Python内置但显式声明更好
